Separate out GHCSpecter.Control.Types
